Initial version of FiddleAndTweak Critic extension
Adds support for editing selected lines via web front-end in the
context of a review, and subsequently pushing the changes to the
review as a follow-up commit.

Lines to edit can be selected as when creating a comment; the
"Create Comment" dialog has an extra button labelled "Edit Code"
which closes the comment creation dialog and opens a code editing
dialog instead.

The lines referenced by an existing comment can also be edited
using the "Edit Code" button in the dialog displaying a comment
thread.